requirements.txt
pandas==1.1.3
numpy==1.19.3
google_trans_new==1.1.9
pyodbc==4.0.30
tensorflow==2.4.1
joblib==0.17.0
streamlit==0.73.1
azure_storage==0.37.0
